+++ empire/server/core/plugin_service.py
@@ -5,7 +5,9 @@ import logging
 import os
 import warnings
 from datetime import datetime
+from pathlib import Path
 
+import yaml
 from sqlalchemy import and_, func, or_
 from sqlalchemy.orm import Session, joinedload, undefer
 
@@ -44,24 +46,23 @@ class PluginService:
         """
         Autorun plugin commands at server startup.
         """
-        plugins = empire_config.yaml.get("plugins")
-        if plugins:
-            for plugin in plugins:
-                use_plugin = self.loaded_plugins.get(plugin)
-                if not use_plugin:
-                    log.error(f"Plugin {plugin} not found.")
-                    continue
+        plugins = empire_config.yaml.get("plugins", {})
 
-                options = plugins[plugin]
-                req = PluginExecutePostRequest(options=options)
+        for plugin_name, options in plugins.items():
+            use_plugin = self.loaded_plugins.get(plugin_name)
+            if not use_plugin:
+                log.error(f"Plugin {plugin_name} not found.")
+                continue
 
-                with SessionLocal.begin() as db:
-                    results, err = self.execute_plugin(db, use_plugin, req, None)
+            req = PluginExecutePostRequest(options=options)
 
-                if results is False:
-                    log.error(f"Plugin failed to run: {plugin}")
-                else:
-                    log.info(f"Plugin {plugin} ran successfully!")
+            with SessionLocal.begin() as db:
+                results, err = self.execute_plugin(db, use_plugin, req, None)
+
+            if results is False:
+                log.error(f"Plugin failed to run: {plugin_name}")
+            else:
+                log.info(f"Plugin {plugin_name} ran successfully!")
 
     def startup_plugins(self, db: Session):
         """
@@ -70,7 +71,47 @@ class PluginService:
         plugin_path = f"{self.main_menu.installPath}/plugins/"
         log.info(f"Searching for plugins at {plugin_path}")
 
-        # Import old v1 plugins (remove in 5.0)
+        self._legacy_load(plugin_path)
+
+        for directory in os.listdir(plugin_path):
+            plugin_dir = Path(plugin_path) / directory
+
+            if (
+                directory == "example"
+                or not plugin_dir.is_dir()
+                or plugin_dir.name.startswith(".")
+                or plugin_dir.name.startswith("_")
+            ):
+                continue
+
+            plugin_yaml = plugin_dir / "plugin.yaml"
+
+            if not plugin_yaml.exists():
+                log.warning(f"Plugin {plugin_dir.name} does not have a plugin.yaml")
+                continue
+
+            plugin_config = yaml.safe_load(plugin_yaml.read_text())
+            plugin_main = plugin_config.get("main")
+            plugin_file = plugin_dir / plugin_main
+
+            if not plugin_file.is_file():
+                log.warning(f"Plugin {plugin_dir.name} does not have a valid main file")
+                continue
+
+            try:
+                self.load_plugin(plugin_file.name.removesuffix(".py"), plugin_file)
+            except Exception as e:
+                log.error(
+                    f"Failed to load plugin {plugin_file.name}: {e}", exc_info=True
+                )
+
+    def _legacy_load(self, plugin_path):
+        warnings.warn(
+            "Legacy plugin loading will be removed in 6.0",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+
         plugin_names = os.listdir(plugin_path)
         for plugin_name in plugin_names:
             if not plugin_name.lower().startswith(
@@ -93,12 +134,16 @@ class PluginService:
                 ):
                     continue
 
+                warnings.warn(
+                    f"{plugin_name}: Loading plugins from .plugin files is deprecated",
+                    DeprecationWarning,
+                    stacklevel=2,
+                )
                 self.load_plugin(plugin_name, file_path)
 
     def load_plugin(self, plugin_name, file_path):
         """Given the name of a plugin and a menu object, load it into the menu"""
-        # note the 'plugins' package so the loader can find our plugin
-        loader = importlib.machinery.SourceFileLoader(plugin_name, file_path)
+        loader = importlib.machinery.SourceFileLoader(plugin_name, str(file_path))
         module = loader.load_module()
         plugin_obj = module.Plugin(self.main_menu)
 
@@ -108,6 +153,7 @@ class PluginService:
             if value.get("Strict") is None:
                 value["Strict"] = False
 
+        plugin_name = plugin_obj.info["Name"]
         self.loaded_plugins[plugin_name] = plugin_obj
 
     def execute_plugin(
@@ -134,9 +180,7 @@ class PluginService:
             log.warning(
                 f"Plugin {plugin.info.get('Name')} does not support db session or user_id, falling back to old method"
             )
-        except PluginValidationException as e:
-            raise e
-        except PluginExecutionException as e:
+        except (PluginValidationException, PluginExecutionException) as e:
             raise e
         except Exception as e:
             log.error(f"Plugin {plugin.info['Name']} failed to run: {e}", exc_info=True)
@@ -153,9 +197,7 @@ class PluginService:
                 )
                 return res
             return res, None
-        except PluginValidationException as e:
-            raise e
-        except PluginExecutionException as e:
+        except (PluginValidationException, PluginExecutionException) as e:
             raise e
         except Exception as e:
             log.error(f"Plugin {plugin.info['Name']} failed to run: {e}", exc_info=True)